+++ lib/regexp_parser/syntax/token/escape.rb
@@ -1,6 +1,6 @@
 module Regexp::Syntax
   module Token
-    # TODO: unify naming with RE::EscapeSequence, on way or the other, in v3.0.0
+    # TODO: unify naming with RE::EscapeSequence, one way or the other, in v3.0.0
     module Escape
       Basic = %i[backslash literal]